pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-intel"
version = "0.1.0"
description = "A website intelligence system that crawls, understands, and answers questions about websites"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Web Intel Team" }]
keywords = ["web-scraping", "llm", "question-answering", "crawler"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Configuration
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    
    # Browser automation
    "playwright>=1.40.0",
    
    # HTTP client
    "httpx>=0.25.0",
    
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # LLM - Local
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    
    # LLM - API
    "anthropic>=0.18.0",
    
    # Embeddings
    "sentence-transformers>=2.2.0",
    
    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
finetune = [
    "peft>=0.7.0",
    "datasets>=2.15.0",
]

[project.scripts]
web-intel = "web_intel.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
